Refactor discovery
