+++ src/usb.rs
@@ -2,16 +2,10 @@ use crate::hw_definition::description::{HardwareDescription, SsidSpec, WiFiDetai
 use crate::hw_definition::usb_values::{
     GET_HARDWARE_VALUE, GET_WIFI_VALUE, PIGGUI_REQUEST, RESET_SSID_VALUE, SET_SSID_VALUE,
 };
-use crate::views::hardware_menu::DeviceEvent;
-use crate::views::hardware_menu::DiscoveryMethod::USBRaw;
-use async_std::prelude::Stream;
-use futures::SinkExt;
-use iced_futures::stream;
 use nusb::transfer::{ControlIn, ControlOut, ControlType, Recipient};
 use nusb::Interface;
 use serde::Deserialize;
 use std::collections::HashMap;
-use std::time::Duration;
 
 /// [ControlIn] "command" to request the HardwareDescription
 const GET_HARDWARE_DESCRIPTION: ControlIn = ControlIn {
@@ -46,7 +40,7 @@ const RESET_SSID: ControlOut = ControlOut {
 /// Try and find an attached "porky" USB devices based on the vendor id and product id
 /// Return a hashmap of interfaces for each one, with the serial_number as the key, enabling
 /// us later to communicate with a specific device using its serial number
-async fn find_porkys() -> HashMap<String, Interface> {
+pub async fn find_porkys() -> HashMap<String, Interface> {
     match nusb::list_devices() {
         Ok(device_list) => {
             let mut map = HashMap::<String, Interface>::new();
@@ -91,16 +85,16 @@ where
 }
 
 /// Request [HardwareDescription] from compatible porky device over USB
-async fn get_hardware_description(porky: &Interface) -> Result<HardwareDescription, String> {
+pub async fn get_hardware_description(porky: &Interface) -> Result<HardwareDescription, String> {
     usb_get_porky(porky, GET_HARDWARE_DESCRIPTION).await
 }
 
 /// Request [WiFiDetails] from compatible porky device over USB
-async fn get_wifi_details(porky: &Interface) -> Result<WiFiDetails, String> {
+pub async fn get_wifi_details(porky: &Interface) -> Result<WiFiDetails, String> {
     usb_get_porky(porky, GET_WIFI_DETAILS).await
 }
 
-/// Send a new Wi-Fi SsidSpec to the connected porky device
+/// Send a new Wi-Fi SsidSpec to the connected porky device over USB
 pub async fn send_ssid_spec(serial_number: String, ssid_spec: SsidSpec) -> Result<(), String> {
     let porkys = find_porkys().await;
     let porky = porkys
@@ -129,63 +123,6 @@ pub async fn reset_ssid_spec(serial_number: String) -> Result<(), String> {
         .get(&serial_number)
         .ok_or("Cannot find USB attached porky with matching serial number".to_string())?;
     usb_send_porky(porky, RESET_SSID).await
-}
-
-/// A stream of [DeviceEvent] announcing the discovery or loss of devices
-pub fn subscribe() -> impl Stream<Item = DeviceEvent> {
-    stream::channel(100, move |gui_sender| async move {
-        let mut previous_serials: Vec<String> = vec![];
-
-        loop {
-            let mut gui_sender_clone = gui_sender.clone();
-            let current_porkys = find_porkys().await;
-
-            // New devices
-            for (serial, porky) in &current_porkys {
-                if !previous_serials.contains(serial) {
-                    match get_hardware_description(porky).await {
-                        Ok(hardware_description) => {
-                            let wifi_details = if hardware_description.details.wifi {
-                                match get_wifi_details(porky).await {
-                                    Ok(details) => Some(details),
-                                    Err(_) => {
-                                        // TODO report error to UI
-                                        None
-                                    }
-                                }
-                            } else {
-                                None
-                            };
-
-                            println!("Found new device");
-                            let _ = gui_sender_clone
-                                .send(DeviceEvent::DeviceFound(
-                                    USBRaw,
-                                    hardware_description.clone(),
-                                    wifi_details,
-                                ))
-                                .await;
-                        }
-                        Err(e) => {
-                            let _ = gui_sender_clone.send(DeviceEvent::Error(e)).await;
-                        }
-                    }
-                }
-            }
-
-            // Lost devices
-            for serial in previous_serials {
-                if !current_porkys.contains_key(&serial) {
-                    let _ = gui_sender_clone
-                        .send(DeviceEvent::DeviceLost(serial.clone()))
-                        .await;
-                }
-            }
-
-            previous_serials = current_porkys.into_keys().collect();
-            tokio::time::sleep(Duration::from_secs(1)).await;
-        }
-    })
 }
 
 /*
