+++ src/views/ssid_dialog.rs
@@ -5,7 +5,7 @@ use self::SsidDialogMessage::{
 
 use crate::widgets::spinner::circular::Circular;
 use crate::widgets::spinner::easing::EMPHASIZED_ACCELERATE;
-use crate::{usb_raw, Message};
+use crate::{usb, Message};
 use iced::keyboard::key;
 use iced::widget::{
     self, checkbox, column, container, horizontal_space, pick_list, row, text, text_input, Button,
@@ -57,7 +57,7 @@ pub enum SsidDialogMessage {
 #[allow(unused_variables)]
 fn send_ssid(serial_number: String, ssid_spec: SsidSpec) -> Task<Message> {
     #[cfg(feature = "usb")]
-    return Task::perform(usb_raw::send_ssid_spec(serial_number, ssid_spec), |res| {
+    return Task::perform(usb::send_ssid_spec(serial_number, ssid_spec), |res| {
         Message::SsidSpecSent(res)
     });
     #[cfg(not(feature = "usb"))]
